+++ packages/spacecat-shared-project-engine-client/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [@adobe/spacecat-shared-project-engine-client-v1.1.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-project-engine-client-v1.1.0...@adobe/spacecat-shared-project-engine-client-v1.1.1) (2026-06-23)
+
+### Bug Fixes
+
+* **project-engine-client:** correct workspace-id param location (CR4) + publish to npm ([#1706](https://github.com/adobe/spacecat-shared/issues/1706)) ([50f52dc](https://github.com/adobe/spacecat-shared/commit/50f52dcb6ab40c9c551624d77c7a17b34e092250))
+
+## [@adobe/spacecat-shared-project-engine-client-v1.1.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-project-engine-client-v1.0.0...@adobe/spacecat-shared-project-engine-client-v1.1.0) (2026-06-23)
+
+### Features
+
+* **project-engine-client:** typed client wrapper + IMS Bearer auth + retry + spec-correction overlay (LLMO-5461) ([#1661](https://github.com/adobe/spacecat-shared/issues/1661)) ([0465ab0](https://github.com/adobe/spacecat-shared/commit/0465ab0e2897a32b64f377241750c8ea0762f0e7))
+
 ## @adobe/spacecat-shared-project-engine-client-v1.0.0 (2026-06-12)
 
 ### Features